requirements.txt
pyodbc>=5.0.0
requests>=2.31.0
google-genai>=0.2.0
python-decouple>=3.8
pandas>=2.1.0
xgboost>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
joblib>=1.3.0

